fix: Modalの名称変更、別ファイルへの分割
