+++ src/components/keywordScreen/index.tsx
@@ -1,5 +1,5 @@
 import { saveKeyword } from '@/lib/memoKeywords';
-import { Box, Button, Center, Text, VStack, ModalOverlay, ModalContent, ModalHeader, ModalCloseButton, ModalBody, ModalFooter, Modal} from '@chakra-ui/react';
+import { Box, Button, Center, Text, VStack } from '@chakra-ui/react';
 import type { FC } from 'react';
 import { useEffect } from 'react';
 import { motion } from 'framer-motion';
@@ -9,27 +9,6 @@ interface KeywordScreenProps {
     route: string
     keyword: string
 }
-interface ModalsProps {
-    errorFlag: boolean
-    onMordalClose: () => void
-}
-
-export const Modals: FC<ModalsProps> = ({ errorFlag, onMordalClose }) => {
-  return (errorFlag&&
-  <Modal isCentered isOpen={errorFlag} onClose={onMordalClose}>
-    <ModalOverlay />
-    <ModalContent>
-      <ModalHeader>Error!</ModalHeader>
-      <ModalCloseButton />
-      <ModalBody>
-            ページが存在していません
-      </ModalBody>
-      <ModalFooter>
-        <Button onClick={onMordalClose}>閉じる</Button>
-      </ModalFooter>
-    </ModalContent>
-  </Modal>);
-};
 
 export const KeywordScreen: FC<KeywordScreenProps> = ({ route, keyword }) => {
   const handleSave = () => {
